+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.92.1
 
+- Move receipt credential expiration check into libsignal